Update chap_00006.txt
Just some shortening of the phrases to help it look more modern and less painful to read, (yes it may be less true to the actual korean, but as long as the meaning of the sentence remains the same, i don't think it should affect it too much). Also while reading i noticed a typo somewhere but i couldn't find it while rereading.