Update chap_00019.txt
Changed: Kwajijijik (과지지직?) is probably an electricity or crackling sound effect. Susuk (스슥) can be the sound of movement or something passing by depending on context, and I interpreted it as the gesture of Mr Goblin swiftly obliterating the food out of existence. Not an ideal translation as I'm not Korean and I don't have the raws to reference, but less confusing than before.